Remove custom directory list routines and replace with a listdir cache to improve compatibility
Changes include
  * Remove custom list directory methods - leave these to pyftpdlib
  * Add ListDirCache to do a listdir using operations.connection.get_container and cache it for 10 seconds
  * Use this in RackspaceCloudFilesFS.listdir
  * Modify RackspaceCloudFilesFS.stat to use the listdir cache
  * Make RackspaceCloudFilesFS.stat for for directories/containers

Not overriding the list methods makes it much more compatible.  It works properly with ncftp now for instance (which uses MSLX).

Putting the ListDirCache in means makes it only use one transaction to list the container with statistics instead of two or many.

ListDirCache has a fallback for stat if a listdir hasn't been cached.