Display the right size/hash for manifest files
Objects larger than 5GB (default) are split in different parts and a special manifest
file is used to access them as a single file.

As getting information of a manifest file requires accessing all the parts, it's an
expensive operation that is not performed in listings, and the manifest files are
listed with 0 bytes and the hash of an empty file.

This change tries to detect manifest files (file size is 0, has a hash and at least is
not a application/directory) and perform an extra request to get the real information of
the manifest file.

References:

 - http://docs.openstack.org/api/openstack-object-storage/1.0/content/large-object-creation.html
 - https://github.com/chmouel/ftp-cloudfs/issues/39